[FE] 불필요한 컴포는트 삭제 및 가독성을 고려한 FeedScrapPanel 코드 수정
