+++ FE/src/styles/mypage/Mypage.module.scss
@@ -234,6 +234,23 @@
   }
 }
 
+.panelWrapper {
+  width: 393px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+
+  .contentWrapper {
+    width: 350px;
+    margin: 30px 0;
+
+    .itemWrapper {
+      margin: 10px 0;
+    }
+  }
+}
+
 .none {
   margin: 40px 40px;
   text-align: center;
